Cleanup productions & legacy parsers
The producers are now more stand-alone. Methods in the lineDataParser
that were only used by a single production are now all part of the production
itself.

As a result I could completly remove the linedata parser
