+++ packages/guides-restructured-text/src/RestructuredText/Parser/Productions/LinkRule.php
@@ -16,6 +16,7 @@ namespace phpDocumentor\Guides\RestructuredText\Parser\Productions;
 use InvalidArgumentException;
 use phpDocumentor\Guides\MarkupLanguageParser;
 use phpDocumentor\Guides\Nodes\AnchorNode;
+use phpDocumentor\Guides\Nodes\Links\Link;
 use phpDocumentor\Guides\Nodes\Links\Link as LinkParser;
 use phpDocumentor\Guides\Nodes\Node;
 use phpDocumentor\Guides\RestructuredText\Parser\DocumentParserContext;
@@ -27,16 +28,9 @@ use phpDocumentor\Guides\RestructuredText\Parser\LinesIterator;
  */
 final class LinkRule implements Rule
 {
-    private LineDataParser $lineDataParser;
-
-    public function __construct(LineDataParser $lineDataParser)
-    {
-        $this->lineDataParser = $lineDataParser;
-    }
-
     public function applies(DocumentParserContext $documentParser): bool
     {
-        $link = $this->lineDataParser->parseLink($documentParser->getDocumentIterator()->current());
+        $link = $this->parseLink($documentParser->getDocumentIterator()->current());
 
         return $link !== null;
     }
@@ -44,7 +38,7 @@ final class LinkRule implements Rule
     public function apply(DocumentParserContext $documentParserContext, ?Node $on = null): ?Node
     {
         $documentIterator = $documentParserContext->getDocumentIterator();
-        $link = $this->lineDataParser->parseLink($documentIterator->current());
+        $link = $this->parseLink($documentIterator->current());
         if ($link === null) {
             throw new InvalidArgumentException();
         }
@@ -57,5 +51,45 @@ final class LinkRule implements Rule
         $documentParserContext->getContext()->setLink($link->getName(), $link->getUrl());
 
         return $node;
+    }
+
+    private function parseLink(string $line): ?Link
+    {
+        // Links
+        if (preg_match('/^\.\. _`(.+)`: (.+)$/mUsi', $line, $match) > 0) {
+            return $this->createLink($match[1], $match[2], Link::TYPE_LINK);
+        }
+
+        // anonymous links
+        if (preg_match('/^\.\. _(.+): (.+)$/mUsi', $line, $match) > 0) {
+            return $this->createLink($match[1], $match[2], Link::TYPE_LINK);
+        }
+
+        // Short anonymous links
+        if (preg_match('/^__ (.+)$/mUsi', trim($line), $match) > 0) {
+            $url = $match[1];
+
+            return $this->createLink('_', $url, Link::TYPE_LINK);
+        }
+
+        // Anchor links - ".. _`anchor-link`:"
+        if (preg_match('/^\.\. _`(.+)`:$/mUsi', trim($line), $match) > 0) {
+            $anchor = $match[1];
+
+            return new Link($anchor, '#' . $anchor, Link::TYPE_ANCHOR);
+        }
+
+        if (preg_match('/^\.\. _(.+):$/mUsi', trim($line), $match) > 0) {
+            $anchor = $match[1];
+
+            return $this->createLink($anchor, '#' . $anchor, Link::TYPE_ANCHOR);
+        }
+
+        return null;
+    }
+
+    private function createLink(string $name, string $url, string $type): Link
+    {
+        return new Link($name, $url, $type);
     }
 }
