+++ packages/guides-restructured-text/src/RestructuredText/Parser/Productions/RuleContainer.php
@@ -1,0 +1,56 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * This file is part of phpDocumentor.
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ *
+ * @link https://phpdoc.org
+ */
+
+namespace phpDocumentor\Guides\RestructuredText\Parser\Productions;
+
+use phpDocumentor\Guides\Nodes\Node;
+use phpDocumentor\Guides\RestructuredText\Parser\DocumentParserContext;
+
+final class RuleContainer
+{
+    /** @var Rule[] */
+    private array $productions;
+
+    public function __construct(Rule ...$productions)
+    {
+        $this->productions = $productions;
+    }
+
+    public function apply(DocumentParserContext $documentParserContext, Node $on): void
+    {
+        $documentIterator = $documentParserContext->getDocumentIterator();
+
+        foreach ($this->productions as $production) {
+            if (!$production->applies($documentParserContext)) {
+                continue;
+            }
+
+            $newNode = $production->apply($documentParserContext, $on);
+            if ($newNode !== null) {
+                $on->addChildNode($newNode);
+            }
+
+            break;
+        }
+
+        $documentIterator->next();
+    }
+
+    public function merge(RuleContainer $productions): self
+    {
+        return new self(
+            ...$this->productions,
+            ...$productions->productions
+        );
+    }
+}
