+++ packages/guides-restructured-text/tests/unit/Parser/Productions/SectionRuleTest.php
@@ -2,6 +2,15 @@
 
 declare(strict_types=1);
 
+/**
+ * This file is part of phpDocumentor.
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ *
+ * @link https://phpdoc.org
+ */
+
 namespace phpDocumentor\Guides\RestructuredText\Parser\Productions;
 
 use phpDocumentor\Guides\RestructuredText\MarkupLanguageParser;
@@ -37,7 +46,7 @@ RST;
             $spanParser->reveal()
         );
 
-        $rule = new SectionRule($titleRule, []);
+        $rule = new SectionRule($titleRule, new RuleContainer());
 
         $document = new DocumentNode('foo', 'index');
 
@@ -66,7 +75,7 @@ RST;
             $spanParser->reveal()
         );
 
-        $rule = new SectionRule($titleRule, []);
+        $rule = new SectionRule($titleRule, new RuleContainer());
 
         $document = new DocumentNode('foo', 'index');
 
@@ -102,7 +111,7 @@ RST;
             $spanParser->reveal()
         );
 
-        $rule = new SectionRule($titleRule, []);
+        $rule = new SectionRule($titleRule, new RuleContainer());
 
         $document = new DocumentNode('foo', 'index');
 
@@ -151,7 +160,7 @@ RST;
             $spanParser->reveal()
         );
 
-        $rule = new SectionRule($titleRule, []);
+        $rule = new SectionRule($titleRule, new RuleContainer());
 
         $document = new DocumentNode('foo', 'index');
